% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitmat3.R
\name{fitmat3}
\alias{fitmat3}
\title{Fit a Matern-III marked point process model via MCMC sampling}
\usage{
fitmat3(mat3, fname = c("centers.txt", "fingers.txt"), win = owin(c(0, 1),
  c(0, 1)), R_clusters = 0.005, R_centers = 0.02, L = 10000, N = 10000,
  J = 100, seed = NULL, resultsName = NULL, logPriors = preparePriors(),
  initialValues = pickInitialValues(), GeyerThompson = c("large",
  "resample", "importance", "simple"), candidateVar = c(0.1, 0.2, 0.1, 0.1,
  0.1), verbose = TRUE, ...)
}
\arguments{
\item{mat3}{A mat3 object, see \code{\link{rmat3}}. If reading from a data file,
set to NA instead.}

\item{fname}{Will only be read if mat3 is set to NA. File names with the centers
and fingers dataset, to be read with \code{\link{read.table}}. The default
action assumes the files "centers.txt" and "fingers.txt", each formatted
with a header and three columns corresponding to a center ID (integer),
and the spatial location of the center/finger end. The fingers file should
have at least one finger for each center ID in the centers file.}

\item{win}{Square window in which the process is to be sampled. See \code{\link{owin}}.
Defaults to the unit square.}

\item{R_clusters}{Nuisance parameter for the fingers' inhibition. No finger endings
will be closer than R_clusters in Euclidean distance during the process
sampling step (hard core inhibiition).}

\item{R_centers}{Nuisance parameter for the centers' inhibition. At the birth-and-death
process, process will take into account how many fingers (attached to other
active centers) are currently at R_centers's distance from the candidate
center being born.}

\item{L}{Number of samples for the MCMC sampler. Defaults to 10000.}

\item{N}{Number of samples for integrating the area of the shadow. Defaults to 10000.}

\item{J}{Number of updates for the Geyer-Thompson step in estimating the sampling
distribution for beta and phi (if GeyerThompson is set to "large", 100 times J).
Defaults to 100 for "simple" and "importance" options in GeyerThompson,
and 10000 for "large".}

\item{seed}{Fixes the RNG seed for replication. Defaults to NULL, which does not
fix the seed.}

\item{resultsName}{Writes the results to a file; defaults to NULL, which does not
save the results.}

\item{logPriors}{A list of 5 log-priors for the coefficients, each being a function of the
parameter and initial value only. See \code{\link{preparePriors}}, which
produces the default values of logPriors.}

\item{initialValues}{Some holistic initial values for empirical priors on the parameters.
See \code{\link{pickInitialValues}}, which produces the default values of
initialValues.}

\item{GeyerThompson}{Provides different methods for updating the Geyer and Thompson (1999)
step in the computation of the likelihood. The "large" option produces J*100
samples from which the Geyer-Thompson approximation ot the likelihood is
built; "importance" uses J samples, but they are updated at every 100
iterations. "simple" works like "importance", but sets phi = 1. "Resample"
uses a resample method on existing finger}

\item{candidateVar}{Variability of MCMC candidate selection step, if tuning is necessary.}

\item{verbose}{Prints acceptance rates for candidates. Defaults to TRUE.}

\item{...}{further arguments passed to \code{\link{read.table}}.}
}
\value{
\code{fitmat3} returns a list containing at least the following components
  \item{parameters}{An L by 5 matrix with the samples from beta, phi, gamma, sigma, kappa.}
}
\description{
Markov chain Monte Carlo sampler from Matern-III marked spatial point process
}
\examples{
set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3)
plot(x)
# Changing default sampling sizes to make it run fast
model <- fitmat3(x, L = 100, N = 1000, J = 2, seed = 1234)
burnin <- 20
layout(matrix(c(1:5,0), ncol=2))
plot(model$parameters[-c(1:burnin),1], type="l", ylab="beta")
abline(h = median(model$parameters[-c(1:burnin),1]))
abline(h = 2.5*25, col = "Red")
plot(model$parameters[-c(1:burnin),2], type="l", ylab="phi")
abline(h = median(model$parameters[-c(1:burnin),2]))
abline(h = 2, col = "Red")
plot(model$parameters[-c(1:burnin),3], type="l", ylab="gamma")
abline(h = median(model$parameters[-c(1:burnin),3]))
abline(h = 20, col = "Red")
plot(model$parameters[-c(1:burnin),4], type="l", ylab="sigma")
abline(h = median(model$parameters[-c(1:burnin),4]))
abline(h = 0.05, col = "Red")
plot(model$parameters[-c(1:burnin),5], type="l", ylab="kappa")
abline(h = median(model$parameters[-c(1:burnin),5]))
abline(h = 3, col = "Red")
# Saving results to file:
\dontrun{model <- fitmat3(NA, fname = c("centers.txt", "fingers.txt"),
                          seed = 1234, resultsName = "results.txt")}

}
\references{
Garcia, N., Guttorp, P. and Ludwig, G. (2018) TBD

  Geyer, C. J., and Thompson, E. A. (1992) "Constrained Monte Carlo maximum likelihood
  for dependent data." Journal of the Royal Statistical Society. Series B (Methodological),
  657-699.
}
\seealso{
\code{\link{rmat3}}
}
\author{
Guilherme Ludwig and Nancy Garcia
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}
