% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitnullmat3.R
\name{fitnullmat3}
\alias{fitnullmat3}
\title{Fit a Poisson marked point process model via MCMC sampling}
\usage{
fitnullmat3(mat3, fname = c("centers.txt", "fingers.txt"),
  win = owin(c(0, 1), c(0, 1)), L = 10000, NN = 2000, seed = NULL,
  resultsName = NULL, logPriors = preparePriors(),
  initialValues = pickInitialValues(), candidateVar = c(0.1, 0.2, 0.1,
  0.1, 0.1), verbose = TRUE, ...)
}
\arguments{
\item{mat3}{A mat3 object, see \code{\link{rmat3}}. If reading from a data file,
set to NA instead.}

\item{fname}{Will only be read if mat3 is set to NA. File names with the centers
and fingers dataset, to be read with \code{\link{read.table}}. The default
action assumes the files "centers.txt" and "fingers.txt", each formatted
with a header and three columns corresponding to a center ID (integer),
and the spatial location of the center/finger end. The fingers file should
have at least one finger for each center ID in the centers file.}

\item{win}{Square window in which the process is to be sampled. See \code{\link{owin}}.
Defaults to the unit square.}

\item{L}{Number of samples for the MCMC sampler. Defaults to 10000.}

\item{NN}{Number of samples to Monte-Carlo estimation of constant. Defaults to 2000.}

\item{seed}{Fixes the RNG seed for replication. Defaults to NULL, which does not
fix the seed.}

\item{resultsName}{Writes the results to a file; defaults to NULL, which does not
save the results.}

\item{logPriors}{A list of 5 log-priors for the coefficients, each being a function of the
parameter and initial value only. See \code{\link{preparePriors}}, which
produces the default values of logPriors.}

\item{initialValues}{Some holistic initial values for empirical priors on the parameters.
See \code{\link{pickInitialValues}}, which produces the default values of
initialValues.}

\item{candidateVar}{Variability of MCMC candidate selection step, if tuning is necessary.}

\item{verbose}{Prints acceptance rates for candidates. Defaults to TRUE.}

\item{...}{further arguments passed to \code{\link{read.table}}.}
}
\value{
\code{fitnullmat3} returns a list containing at least the following components
  \item{parameters}{An L by 5 matrix with the samples from beta, phi, gamma, sigma, kappa.}
}
\description{
Markov chain Monte Carlo sampler from Poisson marked spatial point process
}
\examples{
set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3)
plot(x)
# Changing default sampling sizes to make it run fast
model <- fitnullmat3(x, L = 100, seed = 1234)

}
\references{
Garcia, N., Guttorp, P. and Ludwig, G. (2018) TBD

  Geyer, C. J., and Thompson, E. A. (1992) "Constrained Monte Carlo maximum likelihood
  for dependent data." Journal of the Royal Statistical Society. Series B (Methodological),
  657-699.
}
\seealso{
\code{\link{rmat3}}
}
\author{
Guilherme Ludwig and Nancy Garcia
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}
