% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pagangelouIntensity.R
\name{pagangelouIntensity}
\alias{pagangelouIntensity}
\title{Papangelou Intensity}
\usage{
pagangelouIntensity(xi, beta, phi, fmat, R_c = R_centers, Boot = 1000)
}
\description{
Generates Papangelou conditional intensity function for points in a quadrature.
}
