% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLikelihood.R
\name{plotLikelihood}
\alias{plotLikelihood}
\title{Plot likelihood function}
\usage{
plotLikelihood(final, psi = c(beta = 70, phi = 2, gamma = 5, sigma = 0.05,
  kappa = 3), win = owin(c(0, 1), c(0, 1)), R_clusters = 0.02,
  R_centers = 0.05, J = 100, N = 10000, logPriors = preparePriors(),
  initialValues = pickInitialValues(), GeyerThompson = c("large",
  "importance", "simple"), fname = NULL, logv = TRUE)
}
\arguments{
\item{final}{...}

\item{psi}{...}

\item{win}{...}

\item{R_clusters}{...}

\item{R_centers}{...}

\item{J}{...}

\item{N}{...}

\item{logPriors}{...}

\item{initialValues}{...}

\item{GeyerThompson}{...}

\item{fname}{...}

\item{logv}{...}
}
\description{
Plots the likelihood function for a mat3 object
}
\examples{
\\notrun{x <- rmat3(70, 2, 5, 0.05, 3)
gr <- plotLikelihood(x, GeyerThompson = "importance")
# No importance sampling:
gr <- plotLikelihood(x, psi = c(70, 1, 5, 0.05, 3), GeyerThompson = "simple")
# Reading from a file
gr <- plotLikelihood(NA, fname = c("centers-param=3-b=19.txt", "fingers-param=3-b=19.txt"))}
}
