% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResampledShadow.R
\name{plotResampledShadow}
\alias{plotResampledShadow}
\title{Plot resampled thinned events from shadow}
\usage{
plotResampledShadow(fmat.cond, which.center = 1, which.finger = 1,
  which.proj = c("x", "y"), R_clusters = 0.005, R_centers = 0.02)
}
\description{
This is a debugging function since, by definition, thinned events are not observed in data.
}
\examples{
set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3, returnTimes = TRUE, returnThin = TRUE)
plotResampledShadow(x, 1, 1)
}
