% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictiveStats.R
\name{predictiveStats}
\alias{predictiveStats}
\title{Predictive Analysis for a fitmat3 object}
\usage{
predictiveStats(model, fname = NULL, whichSamples, additionalArgs = list(win
  = owin(), R = 0.005, R_c = 0.02), ...)
}
\arguments{
\item{model}{An fitmat3 object (resulting from a call to fitmat3)}

\item{fname}{If the MCMC samples were saved into text files, this can be used to generate
predictive probabilities from them. Defaults to NULL.}

\item{whichSamples}{Which (and how many) samples to collect in order to generate the
predictive probabilities. Must be smaller than the number of MCMC samples.}

\item{additionalArgs}{additional arguments to be passed to rmat3() (if read from text files,
it is necessary to provide a list with window, R and R_c).}
}
\value{
\code{predictiveStats} returns a list containing at least the following components
  \item{pL}{predictive probability for the expected number of centers}
  \item{pRc}{predictive probability for the number of fingers within R_centers distance of other centers}
  \item{pnx}{predictive probability for the average number of fingers}
  \item{psx}{predictive probability for the standard deviation of the number of fingers}
  \item{chx}{predictive probability for the mean area of the convex hull of fingers (reactive territory)}
  \item{spx}{predictive probability for the spanned angle of fingers}
  \item{mdx}{predictive probability for the average of minimum distance between fingers}
  \item{adx}{predictive probability for the average of average distance between fingers}
}
\description{
Evaluates the predictive posterior probability of T > T_obs for various statistics.
}
