% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudofitmat3.R
\name{pseudofitmat3}
\alias{pseudofitmat3}
\title{Fit a Matern-III marked point process model via MCMC sampling}
\usage{
pseudofitmat3(mat3, fname = c("centers.txt", "fingers.txt"), win = owin(c(0,
  1), c(0, 1)), R_clusters = 0.005, R_centers = 0.02, N = 10000,
  seed = NULL, resultsName = NULL, initialValues = pickInitialValues(),
  ...)
}
\arguments{
\item{mat3}{A mat3 object, see \code{\link{rmat3}}. If reading from a data file,
set to NA instead.}

\item{fname}{Will only be read if mat3 is set to NA. File names with the centers
and fingers dataset, to be read with \code{\link{read.table}}. The default
action assumes the files "centers.txt" and "fingers.txt", each formatted
with a header and three columns corresponding to a center ID (integer),
and the spatial location of the center/finger end. The fingers file should
have at least one finger for each center ID in the centers file.}

\item{win}{Square window in which the process is to be sampled. See \code{\link{owin}}.
Defaults to the unit square.}

\item{R_clusters}{Nuisance parameter for the fingers' inhibition. No finger endings
will be closer than R_clusters in Euclidean distance during the process
sampling step (hard core inhibiition).}

\item{R_centers}{Nuisance parameter for the centers' inhibition. At the birth-and-death
process, process will take into account how many fingers (attached to other
active centers) are currently at R_centers's distance from the candidate
center being born.}

\item{N}{Number of samples for integrating the area of the shadow. Defaults to 10000.}

\item{seed}{Fixes the RNG seed for replication. Defaults to NULL, which does not
fix the seed.}

\item{resultsName}{...}

\item{initialValues}{...}

\item{...}{further arguments passed to \code{\link{read.table}}.}
}
\value{
\code{fitmat3} returns a list containing at least the following components
  \item{parameters}{An L by 5 matrix with the samples from beta, phi, gamma, sigma, kappa.}
}
\description{
Markov chain Monte Carlo sampler from Matern-III marked spatial point process
}
\examples{
set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3)
plot(x)
# Changing default sampling sizes to make it run fast
model <- pseudofitmat3(x, N = 1000, seed = 1234)

}
\references{
Garcia, N., Guttorp, P. and Ludwig, G. (2018) TBD
}
\seealso{
\code{\link{rmat3}}
}
\author{
Guilherme Ludwig and Nancy Garcia
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}
