% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmat3.R
\name{rmat3}
\alias{rmat3}
\title{Sample from Matern-III marked point process}
\usage{
rmat3(beta, phi, gamma, sigma, kappa, win = owin(c(0, 1), c(0, 1)),
  R_clusters = 0.005, R_centers = 0.02, time = 30, death = 1,
  returnTimes = FALSE, formatData = NULL, ...)
}
\arguments{
\item{beta}{Intensity of the dominant Poisson process.}

\item{phi}{Centers' inhibition parameter; rate in which centers are deleted based
on how many fingers from other centers surround it.}

\item{gamma}{Average number of fingers generated.}

\item{sigma}{Length (in standard deviations, according to half-Normal distribution)
of fingers.}

\item{kappa}{Concentration parameter from von Mises distribution.}

\item{win}{Square window in which the process is to be sampled. See \code{\link{owin}}.}

\item{R_clusters}{Hardcore radius of inhibition for finger process.}

\item{R_centers}{Radius of inhibition for center process.}

\item{time}{Length of time for which a birth-and-death process will be sampled.
Defaults to 30.}

\item{death}{Rate in which the birth-and-death process centers will die. Defaults
to 1.}

\item{returnTimes}{Whether the data should be returned with birth times, or just
the sampled spatial locations. Defaults to FALSE. Birth times
are assumed unknown in sampled data, so the birth times should
be used for debugging only.}

\item{formatData}{Saves the data as text files in the current directory, with the
corresponding file names. Set to NULL to override.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{rmat3} returns an object of \code{\link{class}} "mat3" containing
at least the following components
  \item{centers}{The spatial point pattern.}
  \item{fingers}{The marks.}
}
\description{
Generates a sample from Matern-III marked spatial point process
}
\examples{
set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3)
plot(x)
z <- sapply(x, function(x) x$centers)
plot(Kest(ppp(z[1,],z[2,])))
plot(envelope(ppp(z[1,],z[2,])))
# Saves data to  files "centers.txt" and "fingers.txt"
\dontrun{set.seed(1234)
x <- rmat3(70, 2, 5, 0.05, 3, formatData = c("centers.txt", "fingers.txt"))}

}
\references{
Garcia, N., Guttorp, P. and Ludwig, G. (2018) TBD
}
\seealso{
\code{\link{fitmat3}}
}
\author{
Guilherme Ludwig and Nancy Garcia
}
\keyword{Analysis}
\keyword{Data}
\keyword{Functional}
\keyword{Spatial}
\keyword{Statistics}
